requirements.txt
logging
uuid
telegram
_collections
mysql.connector